Create Knights.py
Uses propositional logic to asses built-in logic statements and deriving new inferences to add to the knowledge base. The knowledge base is initialized to the general structure of the "Knights and Knaves" logic riddle, and the AI agent adds more inferences to the base on its own based on the given data. 

While this algorithm was initially designed to solve increasingly complex logic problems through the use of knowledge-based AI, I have built on this project as of late to add more user-friendly functionality. Rather than operating off of built-in statements, the program can take user input instead, converting statements into propositional logic for the AI agent to analyze.